+++ react-ui/src/components/Shared/index.js
@@ -1,5 +1,6 @@
 export { default as Bold } from './Bold';
 export { default as Clickable } from './Clickable';
+export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as Img } from './Img';
 export { default as Link } from './Link';
 export { default as Onboarding } from './Onboarding';